Move query tests to separate file
